+++ Sources/Geometria/2D/Hyperplane2.swift
@@ -18,16 +18,8 @@ public typealias Hyperplane2F = Hyperplane2<Vector2F>
 /// into two.
 public typealias Hyperplane2<V: Vector2FloatingPoint> = Hyperplane<V>
 
-public extension Hyperplane2 {
-    /// Returns the intersection of this 2 dimensional hyperplane with another
-    /// hyperplane as a point, or `nil`, if the hyperplanes are parallel or
-    /// [coplanar](https://en.wikipedia.org/wiki/Coplanarity).
-    func intersection(with other: Self) -> Vector? {
-        let line1 = Line(a: point, b: point + normal.leftRotated())
-        let line2 = Line(a: other.point, b: other.point + other.normal.leftRotated())
-
-        return line1.intersection(with: line2)?.point
-    }
+extension Hyperplane2: PlaneIntersectablePlane2Type {
+    
 }
 
 extension Hyperplane2: Convex2Type {