+++ Sources/Geometria/2D/PointNormalPlane2.swift
@@ -9,3 +9,7 @@ public typealias PointNormalPlane2F = PointNormalPlane2<Vector2F>
 /// Typealias for `PointNormalPlane<V>`, where `V` is constrained to
 /// ``Vector2FloatingPoint``.
 public typealias PointNormalPlane2<V: Vector2FloatingPoint> = PointNormalPlane<V>
+
+extension PointNormalPlane2: PlaneIntersectablePlane2Type {
+    
+}