+++ Sources/Geometria/3D/Protocols/Plane/PlaneIntersectablePlane3Type.swift
@@ -4,7 +4,7 @@
 /// A plane-intersectable plane is implicitly a line-intersectable plane as well.
 public protocol PlaneIntersectablePlane3Type: LineIntersectablePlaneType where Vector: Vector3Type {
     /// Returns the intersection of this 3 dimensional plane with another
-    /// plane as a line, or `nil`, if the hyperplanes are parallel or
+    /// plane as a line, or `nil`, if the planes are parallel or
     /// [coplanar](https://en.wikipedia.org/wiki/Coplanarity).
     func intersection<Plane: LineIntersectablePlaneType>(with other: Plane) -> Line<Vector>? where Plane.Vector == Vector
 }
@@ -23,8 +23,8 @@ extension PlaneIntersectablePlane3Type where Vector: Vector3Multiplicative {
         // created above.
         let toPlane = normal.cross(lineDirection)
 
-        // Making a line projection in the direction of the second hyperplane
-        // should result in an intersection point we can place the line at.
+        // Making a line projection in the direction of the second plane should
+        // result in an intersection point we can place the line at.
         let lineToPlane = Line(a: pointOnPlane, b: pointOnPlane + toPlane)
         guard let pointOnPlane = other.intersection(with: lineToPlane) else {
             return nil